fix: move refund privacy checks into rpc policy
