+++ crates/rpc/Cargo.toml
@@ -15,6 +15,7 @@ workspace = true
 tempo-alloy.workspace = true
 tempo-contracts.workspace = true
 tempo-primitives.workspace = true
+tempo-zone-contracts.workspace = true
 
 # alloy
 alloy-consensus.workspace = true
@@ -25,6 +26,7 @@ alloy-provider = { workspace = true, features = ["reqwest"] }
 alloy-rpc-types-eth.workspace = true
 alloy-signer.workspace = true
 alloy-signer-local.workspace = true
+alloy-sol-types.workspace = true
 
 # rpc server
 axum = { workspace = true, features = ["ws"] }
